feat: simplify local storage syncing
